+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 twig extension Change Log
 2.0.4 under development
 -----------------------
 
-- no changes in this release.
+- Enh #11: Added support for `Twig_SimpleFunction` and `Twig_SimpleFilter` when defining functions and filters (quantum13)
 
 
 2.0.3 March 01, 2015